src/services/ai_layout_adviser.py
from transformers import pipeline
from typing import List, Dict
from dataclasses import dataclass

@dataclass
class LayoutSuggestion:
    component_groups: List[List[str]]
    optimal_routes: Dict[str, List[str]]
    spacing_recommendations: Dict[str, float]

class AILayoutAdviser:
    def __init__(self):
        self.model = pipeline(
            "text-generation",
            model="microsoft/visio-layout-optimizer-v1",
            device=0  # GPU accelerated
        )
        
    def analyze_diagram(self, diagram_json: str) -> LayoutSuggestion:
        """Analyze diagram structure and provide AI suggestions"""
        recommendations = self.model(
            f"Analyze this diagram layout and provide optimization suggestions:\n{diagram_json}"
        )
        return self._parse_recommendations(recommendations[0]['generated_text']) 