requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pydantic>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0

# AI and ML
openai>=1.59.6
langchain>=0.0.300
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Visio integration
pywin32>=228
pythoncom>=0.0.1

# UI and system integration
keyboard>=0.13.5
pyperclip>=1.8.2
psutil>=5.8.0
tkinter>=8.6

# Testing
pytest>=7.0.0
pytest-asyncio>=0.16.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Logging and monitoring
structlog>=21.1.0
prometheus-client>=0.11.0

# Development
black>=21.7b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910

python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.24.1
aiofiles>=23.2.1
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.0.3
fastapi-limiter==0.1.5
redis==4.5.5
backoff==2.2.1
python-magic==0.4.27
tenacity>=8.0.0
jina==3.15.0
docarray==0.21.0
firecrawl-sdk==1.2.0
tqdm>=4.65.0
numpy>=1.24.0
pdfplumber==0.10.0
pymupdf==1.23.21
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
azure-cognitiveservices-vision-computervision==0.9.0
pillow==10.3.0
pdf2image==1.17.0 